# perf(admin-events): fix slow mutex in writer