--- a/central/administration/events/datastore/internal/store/store.go
+++ b/central/administration/events/datastore/internal/store/store.go
@@ -15,4 +15,5 @@ type Store interface {
 	GetByQuery(ctx context.Context, query *v1.Query) ([]*storage.AdministrationEvent, error)
 	UpsertMany(ctx context.Context, objs []*storage.AdministrationEvent) error
 	DeleteMany(ctx context.Context, identifiers []string) error
+	GetMany(ctx context.Context, identifiers []string) ([]*storage.AdministrationEvent, []int, error)
 }