## Description

This PR fixes issues seen during CI where a slow mutex was detected in the administration handler when flushing the buffered events.

The main issue is that previously within the `Upsert` method of the writer a database lookup was done for the event unconditionally.

Since the frequency of `Upsert` being called is directly correlated to the log events being emitted, this put a) a huge amount of read operations on the database and b) lead to the `Upsert` methods holding / waiting on the lock for a long time.

Subsequently, when the `Flush` operation was called from the different go routine, this lead to waiting on the lock.

As a solution, the database lookup was moved to the `Flush` method instead, minimizing the read operations by using `GetMany` and only doing those when the buffer is full.

The benchmark results ran locally can be found below, with a comparison of the previous results:
```
goos: darwin
goarch: amd64
pkg: github.com/stackrox/rox/central/administration/events/datastore
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkDatastore_Add_and_Flush
BenchmarkDatastore_Add_and_Flush/add_1000_events_to_the_writer_and_flush_it
BenchmarkDatastore_Add_and_Flush/add_1000_events_to_the_writer_and_flush_it-12         	      50	  15665028 ns/op
BenchmarkDatastore_Add_and_Flush/add_5000_events_to_the_writer_and_flush_it
BenchmarkDatastore_Add_and_Flush/add_5000_events_to_the_writer_and_flush_it-12         	      50	 111444743 ns/op
BenchmarkDatastore_Add_and_Flush/add_10000_events_to_the_writer_and_flush_it
BenchmarkDatastore_Add_and_Flush/add_10000_events_to_the_writer_and_flush_it-12        	      50	 229537995 ns/op
BenchmarkDatastore_Add_and_Flush/add_20000_events_to_the_writer_and_flush_it
BenchmarkDatastore_Add_and_Flush/add_20000_events_to_the_writer_and_flush_it-12        	      50	 478561343 ns/op


goos: darwin
goarch: amd64
pkg: github.com/stackrox/rox/central/administration/events/datastore
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                                                                       │    old.txt     │                new.txt                │
                                                                       │     sec/op     │    sec/op     vs base                 │
Datastore_Add_and_Flush/add_1000_events_to_the_writer_and_flush_it-12     836.92m ± ∞ ¹   15.67m ± ∞ ¹        ~ (p=1.000 n=1) ²
Datastore_Add_and_Flush/add_5000_events_to_the_writer_and_flush_it-12     6038.3m ± ∞ ¹   111.4m ± ∞ ¹        ~ (p=1.000 n=1) ²
Datastore_Add_and_Flush/add_10000_events_to_the_writer_and_flush_it-12   12902.2m ± ∞ ¹   229.5m ± ∞ ¹        ~ (p=1.000 n=1) ²
Datastore_Add_and_Flush/add_20000_events_to_the_writer_and_flush_it-12   27043.1m ± ∞ ¹   478.6m ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                                                     6.480         117.7m        -98.18%
```

This is an overall increase of performance by over 90% percent.

## Checklist
- [ ] Investigated and inspected CI test results
- [ ] Unit test and regression tests added
- [ ] Evaluated and added CHANGELOG entry if required
- [ ] Determined and documented upgrade steps
- [ ] Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))

If any of these don't apply, please comment below.

## Testing Performed

### Here I tell how I validated my change

- see CI passing.

### Reminder for reviewers

In addition to reviewing code here, reviewers **must** also review testing and request further testing in case the
performed one does not seem sufficient. As a reviewer, you must not approve the change until you understand the
performed testing and you are satisfied with it.